LayoutTests/imported/w3c/web-platform-tests/webrtc/datachannel-emptystring-expected.txt


FAIL Can send empty strings across a WebRTC data channel. assert_equals: Empty message not received expected "" but got "done"
Messages exchanged

Sent: [empty string]
Sent: "done"
Received: "done"
